+++ scaffan/image_intensity_rescale.py
@@ -39,20 +39,22 @@ def rescale_intensity_no_limits(img, in_range, out_range=(-0.9, 0.9)):
 class RescaleIntensityPercentile():
     def __init__(self):
         self.percentile_range = None
-        self.sig_range = None
+        self.percentile_map_range = None
         self.sig_slope = None
         self.input_dtype = None
-        pass
+        self.set_parameters()
 
-    def calculate_parameters(self, img, percentile_range=(5, 95), sig_range=(-0.9, 0.9), sig_slope=1):
+    def set_parameters(self, percentile_range=(5, 95), percentile_map_range=(-0.9, 0.9), sig_slope=1):
         self.percentile_range = percentile_range
-        self.percentile_range_values = np.percentile(img, percentile_range)
-        self.sig_range = sig_range
+        self.percentile_map_range = percentile_map_range
         self.sig_slope = sig_slope
+
+    def calculate_intensity_dependent_parameters(self, img):
+        self.percentile_range_values = np.percentile(img, self.percentile_range)
         self.input_dtype = img.dtype
 
     def rescale_intensity(self, img):
-        imgout = rescale_intensity_no_limits(img, in_range=self.percentile_range_values, out_range=self.sig_range)
+        imgout = rescale_intensity_no_limits(img, in_range=self.percentile_range_values, out_range=self.percentile_map_range)
         imgout = sigmoidal(imgout * self.sig_slope)
         if self.input_dtype == np.uint8:
             imgout = imgout * 255